Move findPackRoot to ql.ts
